+++ lib/src/defineRollDataGetter.ts
@@ -12,6 +12,7 @@ interface Rolled {
 export type RollArray = [number, ...unknown[]][]
 
 export type RollData<T extends Rolled> = {
+  // eslint-disable-next-line no-unused-vars
   readonly [P in keyof T['roll']]: RollArray
 };
