--- src/Alchemist/js/createAlchemist.d.ts
@@ -1,5 +1,0 @@
-interface Setup {
-  createAlchemist(town: Town, opts?: any): Alchemist
-}
-
-interface Alchemist extends Building {}
